# Add cryptoutil::xor_keystream().