--- a/src/chacha20.rs
+++ b/src/chacha20.rs
@@ -8,7 +8,7 @@ use std::simd::u32x4;
 
 use buffer::{BufferResult, RefReadBuffer, RefWriteBuffer};
 use symmetriccipher::{Encryptor, Decryptor, SynchronousStreamCipher, SymmetricCipherError};
-use cryptoutil::{read_u32_le, symm_enc_or_dec, write_u32_le};
+use cryptoutil::{read_u32_le, symm_enc_or_dec, write_u32_le, xor_keystream};
 
 #[derive(Copy)]
 struct ChaChaState {
@@ -251,10 +251,7 @@ impl SynchronousStreamCipher for ChaCha20 {
 
             // Process the min(available keystream, remaining input length).
             let count = cmp::min(64 - self.offset, len - i);
-            for j in range(0, count) {
-                // j-th byte of the keystream, starting from offset.
-                output[i + j] = input[i + j] ^ self.output[self.offset + j];
-            }
+            xor_keystream(&mut output[i..i+count], &input[i..i+count], &self.output[self.offset..]);
             i += count;
             self.offset += count;
         }