The log mostly describes this.  I didn't change the block cipher code, though the CTRx8 implementation should get a decent boost out of something like this.  If you would rather have a 64 bit targeted version, I can do that too, though the gains aren't massive.

The new utility routine does exactly what it says on the tin, except
that it processes 4 bytes at a time (and then handles the remainder if
any).

Before: chacha20::bench::chacha20_64k ... bench: 249568 ns/iter (+/- 24243) = 262 MB/s
After: chacha20::bench::chacha20_64k ... bench: 200760 ns/iter (+/- 4004) = 326 MB/s

It's worth noting that this could be even faster depending on the
target hardware eg:
 * 64 bit -> Process 8 bytes at a time (~350 MiB/s).
 * 64 bit, allows unaligned memory access -> transmute() (~370 MiB/s).

But, ARM probably is a good enough reason to use a generic 32 bit
implementation.