@DaGenix 2015-02-05
Yawning, sorry its taken me so long to get to this. I've been quite busy over the last few days.

This is super cool! I wouldn't have expected such large performance gains to be possible just by optimizing the XORing.

My thoughts on a 64-bit version or unaligned versions: It looks like the performance numbers you outline would make these versions nice to have, if they can be implemented without making the code too complex.

Thanks for this!